hello

+++ class and object/result.cpp
@@ -19,7 +19,7 @@ class result{
     void cal()
     {
         total=eco+eng+stat;
-        per=total/3;
+        per=(float)total/3;
     }
     void getter()
     {
